refactor(coworkers): remove admin-guard and update imports for coworker management access
- Deleted the `admin-guard` module and its associated tests to streamline the codebase.
- Updated import statements in various coworker-related routes to reference the new `coworker-management-access` module for access control.
- This change enhances code organization and maintains functionality for coworker management access.
